Refine server templates implementation
